splitter rename and refactor
